+++ zarrs/examples/optional_data_type.rs
@@ -32,10 +32,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     ])?);
 
     let array = ArrayBuilder::new(
-        vec![4, 4],                                    // 4x4 array
-        vec![2, 2],                                    // 2x2 chunks
-        DataType::Optional(Box::new(DataType::UInt8)), // Optional uint8
-        FillValue::from(None::<u8>),                   // Null fill value: [0]
+        vec![4, 4],                  // 4x4 array
+        vec![2, 2],                  // 2x2 chunks
+        DataType::UInt8.optional(),  // Optional uint8
+        FillValue::from(None::<u8>), // Null fill value: [0]
     )
     .dimension_names(["y", "x"].into())
     .array_to_bytes_codec(Arc::new(OptionalCodec::new(mask_codecs, data_codecs)))