+++ zarrs/tests/optional_codec.rs
@@ -27,10 +27,7 @@ fn create_optional_array(data_type: DataType, fill_value: FillValue) -> Array<Me
 fn optional_array_basic_operations() -> Result<(), Box<dyn std::error::Error>> {
     use ndarray::{array, Array2};
 
-    let array = create_optional_array(
-        DataType::Optional(Box::new(DataType::UInt8)),
-        None::<u8>.into(),
-    );
+    let array = create_optional_array(DataType::UInt8.optional(), None::<u8>.into());
 
     // Store different patterns in different chunks
     // Chunk [0,0]: mostly Some values
@@ -138,7 +135,7 @@ fn optional_array_nested_2_level() -> Result<(), Box<dyn std::error::Error>> {
 
     // Test Option<Option<u8>>
     let array = create_optional_array(
-        DataType::Optional(Box::new(DataType::Optional(Box::new(DataType::UInt8)))),
+        DataType::UInt8.optional().optional(),
         None::<Option<u8>>.into(),
     );
 
@@ -174,9 +171,7 @@ fn optional_array_nested_3_level() -> Result<(), Box<dyn std::error::Error>> {
 
     // Test Option<Option<Option<u16>>>
     let array = create_optional_array(
-        DataType::Optional(Box::new(DataType::Optional(Box::new(DataType::Optional(
-            Box::new(DataType::UInt16),
-        ))))),
+        DataType::UInt16.optional().optional().optional(),
         None::<Option<Option<u16>>>.into(),
     );
 
@@ -219,10 +214,7 @@ fn optional_array_with_non_null_fill_value() -> Result<(), Box<dyn std::error::E
     use ndarray::{array, Array2};
 
     // Create an optional array with a non-null fill value
-    let array = create_optional_array(
-        DataType::Optional(Box::new(DataType::UInt8)),
-        FillValue::from(Some(255u8)),
-    );
+    let array = create_optional_array(DataType::UInt8.optional(), FillValue::from(Some(255u8)));
 
     // Store data in one chunk
     let data = array![
@@ -252,10 +244,7 @@ fn optional_array_string() -> Result<(), Box<dyn std::error::Error>> {
     use ndarray::{array, Array2};
 
     // The ArrayBuilder automatically handles optional strings with the correct vlen codec
-    let array = create_optional_array(
-        DataType::Optional(Box::new(DataType::String)),
-        None::<String>.into(),
-    );
+    let array = create_optional_array(DataType::String.optional(), None::<String>.into());
 
     // Store chunk with Option<String>
     let data = array![