+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `OptionalCodec`
 - Add `ArrayBytesOptional`
 - Add `DataType::Optional`
+- Add `DataType::optional()`
 - Implement `Element` for `Option<T>` where `T: Element`
 - Implement `ElementOwned` for `Option<T>` where `T: ElementOwned`
 - Implement `ElementFixedLength` for `Option<T>` where `T: ElementFixedLength`